Fixes clangd/clangd#1797
Do not restrict HeaderSourceSwitch to switching between .h and .cpp.
For templates, the relevant 'source' is the header containing the implementation of the templated functions.

Instead of considering every symbol in the current file as relevant, we restrict it to those that either have the declaration or definition in that file.